Modificato il tipo di ImportoTotale

+++ CompanyService/Aerei/Biglietto.cs
@@ -5,13 +5,13 @@ public class Biglietto
     public int BigliettoId {get; set;}
     public Volo Volo {get; set;}
     public int PostiPrenotati {get; set;}
-    public double ImportoTotale {get; set;}
+    public decimal ImportoTotale {get; set;}
     public DateTime DataAcquisto {get; set;}
 
     public Biglietto(){
         
     }
-    public Biglietto(int bigliettoId, Volo volo, int postiPrenotati, double importoTotale, DateTime dataAcquisto)
+    public Biglietto(int bigliettoId, Volo volo, int postiPrenotati, decimal importoTotale, DateTime dataAcquisto)
     {
         BigliettoId = bigliettoId;
         Volo = volo;
@@ -19,7 +19,7 @@ public class Biglietto
         ImportoTotale = importoTotale;
         DataAcquisto = dataAcquisto;
     }
-    public Biglietto(Volo volo, int postiPrenotati, double importoTotale, DateTime dataAcquisto)
+    public Biglietto(Volo volo, int postiPrenotati, decimal importoTotale, DateTime dataAcquisto)
     {
         Volo = volo;
         PostiPrenotati = postiPrenotati;
